Added arm64 support to ShareFileOrFolder  class

+++ Owncloud iOs Client/Tabs/FileTab/Share/ShareFileOrFolder.m
@@ -488,9 +488,9 @@
         [[AppDelegate sharedCheckHasShareSupport] updateSharesFromServer];
         [self endLoading];
         
-        DLog(@"error.code: %d", error.code);
-        DLog(@"server error: %d", response.statusCode);
-        int code = response.statusCode;
+        DLog(@"error.code: %ld", (long)error.code);
+        DLog(@"server error: %ld", (long)response.statusCode);
+        NSInteger code = response.statusCode;
         
         //Select the correct msg and action for this error
         switch (code) {
